moved unit_conversion function to a seperate file
