+++ src/valenspy/inputconverter_functions.py
@@ -1,8 +1,8 @@
 from pathlib import Path
 from valenspy.cf_checks import is_cf_compliant, cf_status
 from valenspy._utilities import load_yml
-from valenspy._unit_conversions import (
-    convert_all_units_to_CF,
+from valenspy._unit_conversions import convert_all_units_to_CF
+from valenspy.unit_conversion_functions import (
     _determine_time_interval,
     _convert_mm_to_kg_m2s,
 )
